feat: use one swift class for fabric and paper
